+++ plugins/@sanity/presets/src/index.ts
@@ -1,7 +1,4 @@
-import {definePlugin} from 'sanity'
-
-export function presetsComposer() {
-  return definePlugin({
-    name: '@sanity/presets',
-  })()
-}
+export {presets} from './composer'
+export {linkType, LINK_TYPE_NAME} from './presets/link-type'
+export type {LinkTypeConfig} from './presets/link-type'
+export type {PresetResult} from './types'