+++ plugins/@sanity/presets/src/types.ts
@@ -1,0 +1,5 @@
+import type {SchemaTypeDefinition} from 'sanity'
+
+export interface PresetResult {
+  types: SchemaTypeDefinition[]
+}